refactor: split file-reader live and fake implementations
